Moving tests around.
